Changed Default C++ Version to C++11
Changes:
* Added Cmake rule to force the use of C++11 to allow for vector<string> initialization through lists.
* Changed optimization level from 4 to 3 (compiler kept complaining that O3 = O4).
* Changed some global const char arrays to vector<string> objects in game.cpp to allow revised functions to work properly.
* Removed redundant arguments for number of shaders and number of sounds.
* Commented startGameInstance, getWidth, getHeight, getDirectionalLight, getKeystate, getProgramID methods.
* Refactored and reduced more redundant code throughout other files.
* Revised setupBuild.sh script to notify if build errors are present, preventing the program from autorunning when errors are detected. When build errors are present, a nice red line is printed at the bottom of the console informing the user that there were compilation issues.

+++ textLoader.cpp
@@ -88,7 +88,7 @@ int textLib::genText(FT_Face* face){
         texture,
         ivec2((*face)->glyph->bitmap.width, (*face)->glyph->bitmap.rows),
         ivec2((*face)->glyph->bitmap_left, (*face)->glyph->bitmap_top),
-        (*face)->glyph->advance.x
+        (unsigned int)(*face)->glyph->advance.x
     };
     Characters.insert(std::pair<char, Character>(i, character));
   }
@@ -105,7 +105,7 @@ polygon *textLib::buildTextObj(string text, float x, float y, float scale, vec3 
   {
     // Set ch to the current character at c
     Character ch = Characters[*c];
-    
+
 
     float xpos = x + ch.Bearing.x * scale;
     float ypos = y - (ch.Size.y - ch.Bearing.y) * scale;
